refactor: move FunctionToOptimize to models/function_types to fix circular imports
The FunctionToOptimize class was defined in discovery/functions_to_optimize.py
which has many heavy dependencies (API calls, registry, etc). This caused circular
import issues when modules that needed FunctionToOptimize also needed env_utils.

Moving FunctionToOptimize to models/function_types.py (which is designed to be
dependency-free) breaks the circular import chain.
